refactor : infrastructure/jpa 위치를 user service 공통 위치로 변경
