--- com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/auth/infrastructure/jpa/entity/RoleEnum.java
@@ -1,10 +1,0 @@
-package com.sparta.cupeed.user.auth.infrastructure.jpa.entity;
-
-public enum RoleEnum {
-
-	MASTER, HUB, DELIVERY, COMPANY;
-
-	public String getAuthority() {
-		return "ROLE_" + this.name();
-	}
-}
