--- com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/auth/infrastructure/jpa/entity/StatusEnum.java
@@ -1,6 +1,0 @@
-package com.sparta.cupeed.user.auth.infrastructure.jpa.entity;
-
-public enum StatusEnum {
-
-	PENDING, APPROVED, REJECTED
-}
