+++ com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/infrastructure/jpa/audit/CustomAuditAware.java
@@ -1,4 +1,4 @@
-package com.sparta.cupeed.user.auth.infrastructure.jpa.config.audit;
+package com.sparta.cupeed.user.infrastructure.jpa.audit;
 
 import java.util.Optional;
