+++ com.sparta.cupeed.user/src/main/java/com/sparta/cupeed/user/infrastructure/jpa/entity/RoleEnum.java
@@ -1,0 +1,6 @@
+package com.sparta.cupeed.user.infrastructure.jpa.entity;
+
+public enum RoleEnum {
+
+	MASTER, HUB, DELIVERY, COMPANY
+}
